// src/app/models/request/request-carga.ts
export class RequestCarga {
  uidCargaTrabajo: string;
  motivoAnula: string;
  uidUsuarioE: string;
  uidUsuarioC: string;
  uidOficina: string;
  uidGrupo: string;
  uidEstado: string;
  uidContratista: string;
  descContratista: string;
  uidActividad: string;
  cantidadEjecutada: string;
  cantidadCarga: string;
  fechaSedapal: string;
  fechaContratista: string;
  fechaCarga: string;
  fechaInicio?: string;
  fechaFin?: string;
  idPers: string;
  idPerfil: string;
  vnTipoEstado: number;
  comentario: string;
  usuario: string;
  accion?: string;
  vdescripcion: string;
}